requirements.txt
pydantic>=2.0.0
pydantic-settings>=2.0.0
requests>=2.31.0
urllib3>=2.0.0
mysql-connector-python>=8.2.0
flask>=3.0.0
